Replace timer with lastWrite timestamp
q.Q() calls in the same function in the same file that occur within 2s of each
other are grouped under the same header.

	[14:00:36 main.go main.main:122]

I was using a *time.Timer object to track the 2s intervals to determine if a new
header line needed to be printed. This is a complex object that uses a channel
internally. It's simpler and faster to replace the timer object with a timestamp
for tracking the time since the last write. This enables me to remove init().

Also, I moved the logger type and methods to a separate logger file.
I also disabled the scopelint linter. It kinda sucks.
